Move components from "app" to "src"
