+++ apps/helper-app/src/common/utils.ts
