+++ apps/helper-app/src/register-institution-success/register-institution-success.component.scss
