+++ apps/helper-app/src/register-institution/register-institution.component.ts
@@ -4,7 +4,7 @@ import { Router } from '@angular/router';
 import { customErrorCodes, Institution } from '@wir-vs-virus/api-interfaces';
 import { RecaptchaComponent } from 'ng-recaptcha';
 import { first } from 'rxjs/operators';
-import { InstitutionService } from '../../services/institution.service';
+import { InstitutionService } from '../services/institution.service';
 import { phoneRegExp, zipCodeRegExp } from '../common/utils';
 
 @Component({
