+++ apps/helper-app/src/shared/shared.module.ts
@@ -1,7 +1,7 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { MaterialModule } from '../material/material.module';
 import { InfoBoxComponent } from './info-box/info-box.component';
-import { MaterialModule } from '../../material/material.module';
 
 @NgModule({
   declarations: [InfoBoxComponent],
